% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_f.R
\name{check_if_f}
\alias{check_if_f}
\title{Check if objects are functions mapping a plane}
\usage{
check_if_f(f)
}
\arguments{
\item{f}{argument to check}
}
\description{
Functions checks if its argument is a list of maps of a plane. Returns error otherwise.
}
