% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cool_colours.R
\name{cool_colours}
\alias{cool_colours}
\title{Rainbow-coloured fractals}
\usage{
cool_colours(n, d, ...)
}
\arguments{
\item{n}{number of functions making a fractal}

\item{d}{density of a fractal}

\item{...}{other parameters for rainbow function (see ? rainbow)}
}
\description{
Function cool_colours creates a vector of colours of length n^d to put as
a col parameter in plot.IFS
}
