% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIFS.R
\name{createIFS}
\alias{createIFS}
\title{Create IFS object}
\usage{
createIFS(f)
}
\arguments{
\item{f}{the list of maps (contractions) of a plane.}
}
\description{
Function creates an object of IFS class from a list of functions.
}
