% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragon.R
\docType{data}
\name{dragon}
\alias{dragon}
\title{An example how to create plot of Heighway dragon.}
\format{An object of class \code{list} of length 2.}
\usage{
dragon
}
\description{
The dataset contains functions which allow to create a IFS object. The object is used to plot the fractal named Heigway dragon.
}
\examples{
createIFS(dragon)

}
\keyword{datasets}
