% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fern.R
\docType{data}
\name{fern}
\alias{fern}
\title{An example how to create an IFS Barnsley fern.}
\format{An object of class \code{list} of length 4.}
\usage{
fern
}
\description{
The dataset contains functions which allow to create a IFS object. The object is used to plot the fractal named Barnsley fern.
}
\examples{
createIFS(fern)

}
\keyword{datasets}
