% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractal.R
\name{fractal}
\alias{fractal}
\title{Fractal as a set of points}
\usage{
fractal(f, d)
}
\arguments{
\item{f}{list of functions.}

\item{d}{integer, gives the density of a fractal.}
}
\description{
Function returns a matrix, where each column is a point on a plane. The matrix represents a fractal of a given density d.
}
