% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.R
\name{iterate}
\alias{iterate}
\title{Iterating a list of functions}
\usage{
iterate(f, A)
}
\arguments{
\item{f}{the list of functions}

\item{A}{the matrix, where each column is an iterated point.}
}
\description{
Function creates a matrix, where each column is an image of a point from a given set (represented as a matrix) under a function from the given list of functions.
}
