% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIFS.R
\name{plot.IFS}
\alias{plot.IFS}
\title{Plotting a fractal}
\usage{
\method{plot}{IFS}(x, d = 8, pch = ".", ...)
}
\arguments{
\item{x}{object of class "IFS".}

\item{d}{integer. Determines the density to which plot a fractal.}

\item{pch}{specifies symbols to use when plotting points (see ? plot).}

\item{...}{other parameters of a "plot" function.}
}
\description{
Function plots a fractal from an object of class "IFS" with a given density d.
}
