% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake.R
\docType{data}
\name{snowflake}
\alias{snowflake}
\title{An example how to create an IFS Koch snowflake.}
\format{An object of class \code{list} of length 7.}
\usage{
snowflake
}
\description{
The dataset contains functions which allow to create a IFS object. The object is used to plot the fractal named Koch snowflake.
}
\examples{
createIFS(snowflake)

}
\keyword{datasets}
