% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\docType{data}
\name{triangle}
\alias{triangle}
\title{An example how to create plot of Sierpinski's triangle.}
\format{An object of class \code{list} of length 3.}
\usage{
triangle
}
\description{
The dataset contains functions which allow to create a IFS object. The object is used to plot the fractal named Sierpinski's triangle.
}
\examples{
createIFS(triangle)

}
\keyword{datasets}
